requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-aiplatform==1.39.0
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
structlog==24.1.0
prometheus-client==0.19.0
httpx==0.26.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
tenacity==8.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.14
mypy==1.8.0
pre-commit==3.6.0